pyproject.toml
[tool.poetry]
name = "goit-pythonweb-hw-10"
version = "0.1.0"
description = ""
authors = ["Vasyl Vintoniv <126397140+vinvas94@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = { extras = ["standard"], version = "^0.115.0" }
sqlalchemy = "^2.0.40"
alembic = "^1.15.2"
asyncpg = "^0.30.0"
pydantic-settings = "^2.8.1"
python-dotenv = "^1.0.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
email-validator = "^2.1.0.post1"
cloudinary = "^1.38.0"
redis = "^5.0.1"
fastapi-limiter = "^0.1.5"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.3.5"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

